Fix LevelMinimum and JobLevelMinimum

+++ OpenNos.DAL.EF.MySQL/DB/Item.cs
@@ -35,7 +35,7 @@ namespace OpenNos.DAL.EF.MySQL.DB
         public bool IsSoldable { get; set; }
         public bool IsMinilandObject { get; set; }
         public bool IsWarehouse { get; set; }
-        public short LevelMinimum { get; set; }
+        public byte LevelMinimum { get; set; }
         public short DamageMinimum { get; set; }
         public short DamageMaximum { get; set; }
         public short Concentrate { get; set; }
@@ -49,7 +49,7 @@ namespace OpenNos.DAL.EF.MySQL.DB
         public short DefenceDodge { get; set; }
         public short Hp { get; set; }
         public short Mp { get; set; }
-        public short LevelJobMinimum { get; set; }
+        public byte LevelJobMinimum { get; set; }
         public byte MaxCellon { get; set; }
         public byte MaxCellonLvl { get; set; }
         public short FireResistance { get; set; }
